Update producto + notificaciones

+++ UbyTEC_Web/src/app/Components/BusinessManagerView/management-products/management-products.component.ts
@@ -7,6 +7,8 @@ import { MatTableModule, MatTableDataSource } from '@angular/material/table';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { UpdateDialogComponent } from './update-dialog/update-dialog.component';
 import { ProductService, Product } from '../../../Services/Product/product.service';
+import { DeleteDialogComponent } from './delete-dialog/delete-dialog.component';
+import { NotificationService } from '../../../Services/Notification/notification.service';
 
 @Component({
   selector: 'app-management-products',
@@ -19,6 +21,7 @@ export class ManagementProductsComponent {
   constructor(
     private dialog: MatDialog,
     public service: ProductService,
+    public notificationService: NotificationService
   ) {}
 
   listData = new MatTableDataSource<any>
@@ -41,13 +44,26 @@ export class ManagementProductsComponent {
     this.dialog.open(CreateDialogComponent, dialogConfig).afterClosed().subscribe(() => {this.ngOnInit()}); // Actualiza lista luego de cerrar
   }
 
-  updateDialog(): void {
+  updateDialog(code:number, data: any): void {
+    this.service.populateForm(data);
+    const dialogConfig = new MatDialogConfig();
+    dialogConfig.autoFocus = true;
+    this.dialog.open(UpdateDialogComponent, dialogConfig).afterClosed().subscribe(() => {this.ngOnInit()}); // Actualiza lista luego de cerrar
   }
 
   deleteProducto(row : any): void {
-    this.service.delete(row).subscribe((row)=>{
-      this.ngOnInit();
+    this.dialog.open(DeleteDialogComponent, {
+      disableClose: true,
+      data: {
+        message: '¿Eliminar esta entrada?'
+      }
+    }).afterClosed().subscribe(res => {
+      if (res) {
+        this.service.delete(row).subscribe((row)=>{
+          this.ngOnInit();
+          this.notificationService.success('Producto eliminado con éxito');
+        });
+      }
     })
-    
   }
 }
